chore: Follow package rename for jnigen-loader
